Settings: Add preference for toggling the wallpaper zooming animation [2/2]
For some reason Google decided that permanently zooming the wallpaper by 10%
and ruining its' quality in order to show a "zoom out depth animation" was a
good design choice.
Make this user configurable in order to get our pixels back.

With toggle at it's default AOSP value:
https://i.imgur.com/3aGNNij.png

With toggle at off (you can actually see the full image as it's supposed to be):
https://i.imgur.com/GtAiqsN.png